refactor(services): extract web chat state and messaging client
Relocates the chat core - messagesBySession, subscribersBySession,
activeStreams and nextMessageId - into web-chat-state.ts, which owns
them and exports behaviour. No binding is exported.

This is the cut that changes the graph: with those four relocated, the
one remaining component splits into {webAgentMemories,
nextAgentMemoryId} and {memoryChatConfigs, recoveryReportsBySession},
which the following groups can then take independently.

Six methods move into web-chat-client.ts over the new
ChatMessagingService, whose signatures moved out of AgentService rather
than being copied. sendMessage stays in the composition root and now
reaches chat state through the accessors.

resetWebLoopsForTest returns to a single function in web-loop-state.ts,
so clearWebLoopTimersAndSubscribers and clearWebLoopStateForTest stop
being exported half-steps.

web-agent-client.ts 1877 -> 1608, agent-service.ts 364 -> 346.
Subtree aggregate 19087 -> 19164 over three new modules.

tsc --noEmit, contracts:check and the full suite pass: 286 files /
1301 tests, unchanged from baseline, with no test file edited.
