+++ scripts/architecture/frontend-rules.mjs
@@ -13,8 +13,11 @@ import { architectureDiagnostic, architectureSummaryDiagnostic, RULES } from "./
 // 绑定搬进 state 模块时,除了上面那份固定开销,还多出一层访问器——原先一句直接读写(`sessions`、
 // `activeSessionId`、`webSkills`)现在是一对 `list*`/`replace*` 函数。这层是"禁止导出可变绑定"
 // 这条规则的代价,不是重复代码:访问器集合按调用点裁剪,每个都有本模块之外的调用者。
+// 上调理由(extract-web-client-chat-state):同上一条,把最后 8 个共享可变绑定搬进 state 模块的
+// 代价还是那两项——每个新模块的固定开销,加上一层按调用点裁剪的访问器。这一轮多出来的访问器都在
+// chat 一侧:原先直接 `activeStreams.has/set/delete`、`messagesBySession.delete`,现在各是一个函数。
 const SUBTREE_LINE_BUDGETS = Object.freeze([
-  { root: "src/services", budget: 19087, owner: "extract-web-client-state-modules" },
+  { root: "src/services", budget: 19164, owner: "extract-web-client-chat-state" },
 ]);
 
 const STATE_PACKAGES = new Set([
